+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 0.1.0-dev
+### Added
+* Way to specify payment id when we ask a payment
+* Abort payment request
+
 ## 0.0.2-dev
 ### Added
 * Enumeration of available endpoints.